Allow to copy, cut and paste elements across browser's windows (#341)
Copied elements are saved in the system clipboard. Paste button is currently supported only in Chrome like browsers and it requires the user to grant access to the clipboard. Ctrl+V shortcut should work in most browsers. What's more paste button is always enabled since we cannot easily detect when clipboard content was modified.

Relates to #225.