Make FixedValuesEditor more explicit about current value
Problem I've found.
Editor would render the first option when the current value is invalid or blank.
It's misleading, user has no idea that the true value in process (as seen with
JSON export) is different

I've simplified the FixedValuesEditor so that it renders what it gets.
What in case the value is no longer valid?
  Validation should take care and force user to select a new value.
What aboud default value?
  I think it should be set on process/node creation time. It has nothing to do with render time.