NU-1074 Updated FixedValuesGroup component and test
Changed the previously set fixedValuesType and initialValue properties in the FixedValuesGroup component and corresponding test. These changes are implemented for better null handling and to comply with the new structure, where the value type is under a property named "valueEditor.type" instead of "fixedValuesType". This provides more clarity and reduces ambiguity in our codebase.