Update inputMode comment in fragment.cy.ts
Changed the comment in fragment.cy.ts to accurately reflect the provided value inputMode. The comment has been updated from 'String Any with Suggestion Value' to 'String Fixed value' to avoid any possible confusion in reading or interpreting the code. Fully understanding the purpose of each part of the code is crucial for maintaining and expanding the codebase.